refactor: Fare Policy 패키징
