+++ src/main/java/nextstep/subway/domain/policy/fare/OverFiftyExtraFare.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain.policy;
+package nextstep.subway.domain.policy.fare;
 
 import org.springframework.stereotype.Component;
