+++ src/test/java/nextstep/subway/domain/policy/fare/ExtraLineFareTest.java
@@ -1,6 +1,9 @@
-package nextstep.subway.domain.policy;
+package nextstep.subway.domain.policy.fare;
 
 import nextstep.subway.domain.Line;
+import nextstep.subway.domain.policy.fare.ExtraLineFare;
+import nextstep.subway.domain.policy.fare.FarePolicy;
+import nextstep.subway.domain.policy.fare.PathByFare;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
